♻️ Merge all storybook files into a single one
