--- src/app/components/Icons/Friends.stories.tsx
@@ -1,9 +1,0 @@
-import React from 'react';
-import Friends from './Friends';
-
-export default {
-  title: 'Component/Icons',
-  component: Friends,
-};
-
-export const FriendsIcon = (): JSX.Element => <Friends />;
