--- src/app/components/Icons/Items.stories.tsx
@@ -1,9 +1,0 @@
-import React from 'react';
-import Items from './Items';
-
-export default {
-  title: 'Component/Icons',
-  component: Items,
-};
-
-export const ItemsIcon = (): JSX.Element => <Items />;
