--- src/app/components/Icons/Map.stories.tsx
@@ -1,9 +1,0 @@
-import React from 'react';
-import Map from './Map';
-
-export default {
-  title: 'Component/Icons',
-  component: Map,
-};
-
-export const MapIcon = (): JSX.Element => <Map />;
