--- src/app/components/Icons/News.stories.tsx
@@ -1,9 +1,0 @@
-import React from 'react';
-import News from './News';
-
-export default {
-  title: 'Component/Icons',
-  component: News,
-};
-
-export const NewsIcon = (): JSX.Element => <News />;
