--- src/app/components/Icons/Stats.stories.tsx
@@ -1,9 +1,0 @@
-import React from 'react';
-import Stats from './Stats';
-
-export default {
-  title: 'Component/Icons',
-  component: Stats,
-};
-
-export const StatsIcon = (): JSX.Element => <Stats />;
